refactor: decompose admin handler module (#151)
